+++ cgbeacon2/cli/delete.py
@@ -3,7 +3,8 @@
 import click
 from flask.cli import with_appcontext, current_app
 
-from cgbeacon2.utils.delete import delete_dataset
+from cgbeacon2.utils.delete import delete_dataset, delete_variants
+from cgbeacon2.utils.update import update_dataset_samples
 
 
 @click.group()
@@ -24,7 +25,7 @@ def dataset(id):
 
     click.echo(f"deleting dataset with id '{id}' from database")
 
-    deleted = delete_dataset(mongo_db=current_app.db, id=id)
+    deleted = delete_dataset(database=current_app.db, id=id)
 
     if deleted is None:
         click.echo("Aborting")
@@ -32,3 +33,56 @@ def dataset(id):
         click.echo(f"Coundn't find a dataset with id '{id}' in database.")
     elif deleted == 1:
         click.echo("Dataset was successfully deleted")
+
+
+@delete.command()
+@with_appcontext
+@click.option("-ds", type=click.STRING, nargs=1, required=True, help="dataset ID")
+@click.option(
+    "-sample",
+    type=click.STRING,
+    multiple=True,
+    required=True,
+    help="one or more samples to remove variants for",
+)
+def variants(ds, sample):
+    """Remove variants for one or more samples of a dataset
+
+    Accepts:
+        ds(str): id of a dataset already existing in the database
+        sample(str) sample name as it's written in the VCF file, option repeated for each sample
+    """
+
+    click.confirm(
+        f"Deleting variants for sample {sample}, dataset '{ds}'. Do you want to continue?",
+        abort=True,
+    )
+
+    # Make sure dataset exists and contains the provided sample(s)
+    dataset = current_app.db["dataset"].find_one({"_id": ds})
+    if dataset is None:
+        click.echo(f"Couldn't find any dataset with id '{ds}' in the database")
+        raise click.Abort()
+
+    for s in sample:
+        if s not in dataset.get("samples", []):
+            click.echo(
+                f"Couldn't find any sample '{s}' in the sample list of dataset 'dataset'"
+            )
+            raise click.Abort()
+
+    updated, removed = delete_variants(current_app.db, ds, sample)
+    click.echo(f"Number of variants updated:{updated}, removed:{removed}")
+
+    if updated + removed > 0:
+        # remove sample(s) from dataset
+        result = update_dataset_samples(
+            database=current_app.db, dataset_id=ds, samples=list(sample), add=False
+        )
+
+        if result is not None:
+            click.echo(
+                f"Samples {sample} were successfully removed from dataset list of samples"
+            )
+        else:
+            click.echo("List of dataset samples was left unchanged")