+++ cgbeacon2/utils/add.py
@@ -19,8 +19,6 @@ def add_dataset(database, dataset_dict, update=False):
     Returns:
         inserted_id(str): the _id of the added/updated dataset
     """
-
-    inserted_id = None
     collection = "dataset"
 
     if update is True:  # update an existing dataset
@@ -171,7 +169,7 @@ def add_variant(database, variant, dataset_id):
         if len(updated_samples) > 0:
             old_datasets_dict[dataset_id]["samples"] = updated_samples
             # update variant with new samples
-            result = database["variant"].update_one(
+            result = database["variant"].find_one_and_update(
                 {"_id": old_variant["_id"]}, {"$set": {"datasetIds": old_datasets_dict}}
             )
-            return result.modified_count
+            return True