+++ cgbeacon2/utils/update.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import datetime
 import logging
 
 LOG = logging.getLogger(__name__)
@@ -27,11 +28,20 @@ def update_dataset_samples(database, dataset_id, samples, add=True):
     datasets_samples_size = len(datasets_samples)
 
     for sample in samples:  # add new samples to dataset
-        datasets_samples.add(sample)
+        if add is True:
+            datasets_samples.add(sample)
+        else:
+            datasets_samples.remove(sample)
 
-    if len(datasets_samples) > datasets_samples_size:
+    if len(datasets_samples) != datasets_samples_size:
         # update dataset with new samples
         result = database["dataset"].find_one_and_update(
-            {"_id": dataset_id}, {"$set": {"samples": list(datasets_samples)}}
+            {"_id": dataset_id},
+            {
+                "$set": {
+                    "samples": list(datasets_samples),
+                    "updated": datetime.datetime.now(),
+                }
+            },
         )
         return result